debug

+++ src/DSMKeeper.cpp
@@ -256,7 +256,7 @@ uint64_t DSMKeeper::max(const std::string &max_key, double value,
   memSet(key.c_str(), key.size(), (char *)&value, sizeof(value));
 
   uint64_t ret = 0;
-  uint64_t max = 0;
+  double max = 0;
   for (int i = 0; i < node_num; ++i) {
     key = key_prefix + std::to_string(i);
     auto mem_ret = memGet(key.c_str(), key.size(), nullptr, false);
